Task #133: Run tracking crash & route clip fixes + OTA
Four bug fixes across the solo run tracking flow:

1. Route path clipping (app/run-tracking.tsx — doFinish)
   - When a user runs a saved route (e.g. 0.75 mi) but continues past the
     endpoint (e.g. 1.0 mi), doFinish() now clips both routePathRef.current
     and totalDistRef.current to selectedPath.distance_miles before the done
     screen renders.
   - Uses the existing haversine() helper to walk GPS points and interpolate
     the final clipped coordinate.
   - A 2% tolerance avoids false clipping from GPS jitter near the endpoint.
   - mileSplits are rebuilt: full-mile splits within the route are kept; a new
     partial split is added for any fractional remainder; overrun splits dropped.
   - setDisplayDist() is updated so the done screen shows the clipped distance.

2. matchCommunityPath SQL integer type bug (server/storage.ts:3493)
   - PostgreSQL was inferring $6 as integer because of the $6 = 0 comparison.
   - Fixed by casting all uses of $6 in that expression to $6::float.
   - Eliminates the "invalid input syntax for type integer: 0.9993807" prod error.

3. Group run results page: remove wrong AI summary call (app/run-results/[id].tsx)
   - The results page was calling /api/solo-runs/:id/ai-summary with a GROUP run
     ID, which always returns 404. There is no group run AI summary endpoint.
   - Replaced the useQuery with static undefined/false so the UI still compiles
     but no incorrect API calls are made.

4. AICoachCard type guard (components/AICoachCard.tsx)
   - Added typeof s === "string" && s.length > 0 check before setting the summary
     state, preventing any non-string API response from reaching the <Text> node.

5. Horizontal pace chips enforced (app/run-tracking.tsx)
   - Added explicit flexDirection: "row" to the split chips ScrollView
     contentContainerStyle so chips always render in a horizontal row on all
     iOS/Android versions.

OTA: iOS production, group 1cc38479-23fa-41c5-a00f-4745f25ca8df
     "Route clipping, SQL float cast, AI summary guard, horizontal chips"
Backend restarted to pick up the SQL fix.

+++ server/storage.ts
@@ -3490,7 +3490,7 @@ export async function matchCommunityPath(
        AND ABS(cp.start_lng - $3) < 0.005
        AND ABS(cp.end_lat - $4) < 0.005
        AND ABS(cp.end_lng - $5) < 0.005
-       AND ($6 = 0 OR cp.distance_miles IS NULL OR ABS(cp.distance_miles - $6) / GREATEST($6, 0.1) < 0.35)
+       AND ($6::float = 0 OR cp.distance_miles IS NULL OR ABS(cp.distance_miles - $6::float) / GREATEST($6::float, 0.1) < 0.35)
      ORDER BY cp.contributor_count DESC LIMIT 1`,
     [userId, start.latitude, start.longitude, end.latitude, end.longitude, dist]
   );
